+++ src/main/java/com/cburch/logisim/circuit/CircuitAttributes.java
@@ -52,6 +52,7 @@ import java.util.Arrays;
 import java.util.List;
 
 public class CircuitAttributes extends AbstractAttributeSet {
+
   private class MyListener implements AttributeListener, CircuitAppearanceListener {
     private final Circuit source;
 
@@ -161,8 +162,6 @@ public class CircuitAttributes extends AbstractAttributeSet {
       Attributes.forFont("clabelfont", S.getter("circuitLabelFontAttr"));
   public static final Attribute<Boolean> CIRCUIT_IS_VHDL_BOX =
       Attributes.forBoolean("circuitvhdl", S.getter("circuitIsVhdl"));
-  public static final Attribute<String> CIRCUIT_VHDL_PATH =
-      Attributes.forString("circuitvhdlpath", S.getter("circuitVhdlPath"));
   public static final Attribute<Boolean> NAMED_CIRCUIT_BOX_FIXED_SIZE =
       Attributes.forBoolean("circuitnamedboxfixedsize", S.getter("circuitNamedBoxFixedSize"));
   public static final AttributeOption APPEAR_CLASSIC = StdAttr.APPEAR_CLASSIC;
@@ -175,6 +174,8 @@ public class CircuitAttributes extends AbstractAttributeSet {
           "appearance",
           S.getter("circuitAppearanceAttr"),
           new AttributeOption[] {APPEAR_CLASSIC, APPEAR_FPGA, APPEAR_EVOLUTION, APPEAR_CUSTOM});
+  public static final Attribute<Double> SIMULATION_FREQUENCY = Attributes.forDouble("simulationFrequency");
+  public static final Attribute<Double> DOWNLOAD_FREQUENCY = Attributes.forDouble("downloadFrequency");
 
   private static final Attribute<?>[] STATIC_ATTRS = {
     NAME_ATTR,
@@ -183,11 +184,12 @@ public class CircuitAttributes extends AbstractAttributeSet {
     CIRCUIT_LABEL_FONT_ATTR,
     APPEARANCE_ATTR,
     NAMED_CIRCUIT_BOX_FIXED_SIZE,
-    CIRCUIT_VHDL_PATH
+    SIMULATION_FREQUENCY,
+    DOWNLOAD_FREQUENCY,
   };
 
   private static final Object[] STATIC_DEFAULTS = {
-    "", "", Direction.EAST, StdAttr.DEFAULT_LABEL_FONT, APPEAR_CLASSIC, false, ""
+    "", "", Direction.EAST, StdAttr.DEFAULT_LABEL_FONT, APPEAR_CLASSIC, false, -1d, -1d
   };
 
   private static final List<Attribute<?>> INSTANCE_ATTRS =
@@ -201,8 +203,7 @@ public class CircuitAttributes extends AbstractAttributeSet {
           CIRCUIT_LABEL_ATTR,
           CIRCUIT_LABEL_FACING_ATTR,
           CIRCUIT_LABEL_FONT_ATTR,
-          APPEARANCE_ATTR,
-          CIRCUIT_VHDL_PATH);
+          APPEARANCE_ATTR);
 
   private final Circuit source;
   private Instance subcircInstance;
@@ -225,6 +226,8 @@ public class CircuitAttributes extends AbstractAttributeSet {
     LabelVisible = true;
     pinInstances = new Instance[0];
     NameReadOnly = false;
+    DOWNLOAD_FREQUENCY.setHidden(true);
+    SIMULATION_FREQUENCY.setHidden(true);
   }
 
   @Override